\name{bceInput}
\docType{data}
\alias{bceInput}
\title{ratiomatrix and datamatrix for demonstration of BCE()}
\description{The datamatrix contains a set of biomarker measurements for a number of field samples. 
The ratiomatrix contains biomarker data of a number of biological taxa. 
BCE() uses these matrices to estimate the taxonomical composition of the samples based on the provided taxa.}
\usage{bceInput}
\keyword{datasets}

