\name{bceOutput}
\docType{data}
\alias{bceOutput}
\title{bce output generated by running the bceInput example}
\description{Result generated by running BCE using data \code{bceInput} as input.
\cr the run was initiated with the following command:
\cr \code{bceOutput <- BCE(bceInput$Rat,bceInput$Dat,relsdRat=.2,relsdDat=.2,}\cr
\code{iter=100000,outputlength=5000,jmpX=.02,jmpRat=.2*(.2*bceInput$Rat))}
\cr this run took several minutes.
}
\usage{bceOutput}
\examples{

summary(bceOutput)$meanX

}
\keyword{datasets}

