\name{plot.bce}
\alias{plot.bce}
\title{plot BCE}
\description{produces summary plots of the random walks of the BCE; these are intended for inspection only}
\usage{plot.bce(bce)}
\arguments{
  \item{bce }{bce object}
}
\author{Karel Van den Meersche}
\examples{
# bceOutput is an example output based on bceInput

plot(bceOutput)
}

\keyword{models}

